pyproject.toml
[tool.poetry]
name = "church-of-gromp-dodgeball"
version = "0.1.0"
description = ""
authors = ["Parker Freestone <parkerfreestone@gmail.com>"]
readme = "README.md"
packages = [
    { include = "lib" }
]

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
fastapi = {extras = ["standard"], version = "^0.115.2"}
peewee = "^3.17.7"
python-dotenv = "^1.0.1"
aiohttp = "^3.10.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
